requirements.txt

fastapi==0.115.8
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.10.6
langchain==0.3.24
langchain-community==0.3.17
langchain-core==0.3.56
openai==1.74.0
anthropic==0.46.0
google-generativeai==0.8.4
chromadb==0.4.15
sentence-transformers==3.4.1
python-multipart==0.0.20
jinja2==3.1.5
markdown==3.7
ipython==8.32.0
click==8.1.8
uvloop==0.19.0
httpx==0.28.1 
pinecone==3.1.0
tiktoken==0.7.0
requests==2.32.3
sse-starlette>=1.6.5 